+++ src/index.js
@@ -227,6 +227,22 @@ class KeyringController extends EventEmitter {
             return Promise.reject(e)
         }
     }
+    importWallet(_privateKey) {
+        try {
+            if (_privateKey.startsWith('0x')) {
+                _privateKey = _privateKey.slice(2)
+            }
+            const privateKey = Buffer.from(_privateKey, 'hex')
+            if (!ethUtil.isValidPrivate(privateKey))
+                throw "Enter a valid private key"
+
+            const address = ethUtil.bufferToHex(ethUtil.privateToAddress(privateKey))
+            this.importedWallets.push(address);
+            return address
+        } catch (e) {
+            return Promise.reject(e)
+        }
+    }
 
     //
     // SIGNING METHODS